Rework of cache configuration
Cache directory path configuration does no rely on environment variable
anymore. This rework introduces a new interface FileBasedCache.configure
to configure the cache. For now, there are two configurations, cache_dir
and disabled. Config disabled is used for testing to test different
behavior when cache is enabled or not.

Default cache directory is still created if a cache directory path is
not specified from the command line.

Tests are updated accordingly as well as the fixtures.