Retain block sequence indentation
STONEBLD-3422

Users may use various block sequence indentation styles for themselves
that are different from the one proposed by Konflux during the
onboard. This commit tries to detect the indentation styles used in the
users side and retain the style as much as possible in order to not
introduce big formatting change to users. The tool does:

* if block sequence indentation style is used consistently with same
  levels through the whole file, whatever 0, 2, or 4, the result
  pipeline keeps same style.
* if various styles are mixed, the block sequneces are formatted without
  indentation in the result pipeline.

This commit makes these changes mainly:

* Add a new class YAMLStyle for detecting the indentation sytles.
* YAML related methods in utils.py are extended to support dumping data
  with styles.
* Move fixed styles width and preserve_quotes into YAMLStyle class.
* resolve_pipeline method ensures the styles are detected and save the
  pipeline with detected style.
* Tests are added and updated accordingly. The tests involves YAMLs with
  different styles of block sequence indentations, then part of the test
  code are refactored with pytest fixture in order to write tests
  easily.