Decouple method search_pipeline_files
The problem of method search_pipeline_files is it mixes the iteration
over various paths, predication of what files to be handled, and the how
the files are handled. This is good for migrate sub-command for which
the search_pipeline_files method was introduced, but it is not flexible
for other operations (sub-commands) against pipeline files. For example,
not all operations require the behavior of extracting .spec.pipelineSpec
into a separate temporary file for handling a PipelineRun definition.

Major changes:

* Extract iterate_files_or_dirs method, that is dedicated for searching
  YAML files.
* Create a new class PipelineFileOperation to make decision whether a
  file is an expected Pipeline or PipelineRun definition file and should
  be handled. If yes, it dispatches to dedicated handle_* methods.
* PipelineFileOperation must be sub-classed in order to implement
  specific logic.
* migrate and add-task commands are refactored by creating own operation
  classes derived from PipelineFileOperation.
* Tests are refactored and updated according to the above decouple.