fix: rework determine_task_bundle_upgrades_range
This is a further rework based on the previous fix of dropping
out-of-order bundles. This reworks tries to handle new tag scheme made
during transitioning to the decentralization.
determine_task_bundle_upgrades_range has the all detailed information
about this rework.

Please note that this is not the solution for addressing the new tag
scheme of decentralization.

Major changes:

* Method determine_task_bundle_upgrades_range is rewritten based on the
  cleaned up tags by dropping out-of-order bundles.
* Retrieval of tags is reimplemented. It nows queries tags by versions
  specifically.
* Test TestDetermineTaskBundleUpdatesRange is rewritten according to the
  rework and other tests are updated accordingly.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>