List bundle tags by versions explicitly
The major difficulty of determining the upgrade range is to identify the
range start and end. Since bundles can be built from an old version,
even any older ones, the tags responded by Quay.io listRepoTags endpoint
are mixed with different versions through the timeline from newest to
oldes. For possible cases, please refer to the test, which contains
various potential cases as much as possible.

This patch proposes a new mechanism to list tags by versions explicitly
rather than trying to handle different cases by iterating tags linearly.
As a result, irrelevant tags will not be present in the tag list, and it
is much easier to drop out-of-order tags without noise.

This way requires to version expansion according to the old and new
version, for example, upgrade from current value 0.2 to new value 0.4,
it will be expanded to 0.2, 0.3 and 0.4. This expansion relies on the
current task version management strategy used by build-definitions.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>