Auto-switch to MigrationImagesResolver
STONEBLD-3776

LinkedMigrationsResolver is working for migrations pushed by
build-definitions CI. During the transition to decentralized task
repositories, more and more tasks are started to be built by tekton
bundle builder pipeline and the others are still built by
build-definitions CI. So, the root problem here is migrations of every
bundle upgrade should be discovered and fetched by a specific resolver.

This commit introduces a resolver proxy. It selects
MigrationImagesResolver for the decentralized bundles, and
LinkedMigrationsResolver for the others.

It is an auto-switch. The condition is whether a migration image tag is
present in a bundle's image repository. If such tag is not present, it
may mean either there is actually no migrations yet for a decentralized
task, or the task is not decentralized yet. As a result,
LinkedMigrationsResolver may run for the former case. This could be
confused, but it is not a big problem, as no migration will be
discovered obviously.

When to drop the resolver proxy?

Resolver proxy and the related changes in this commit can be removed
once the decentralization transition is fully complete.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>