Refine venv management
Motivation: pip-compile running with python3.13 generates requirements
without including packages that are only required for <=python3.12, that
will cause failure happening to py312 testenv creation.

An example package is ruamel.yaml.clib, that is required by recent
ruamel.yaml releases as of this commit:

    platform_python_implementation=="CPython" and python_version<"3.14"

As Python 3.12 is the minimum version to work with, this commit refines
the Makefile to create a virtual environment with Python 3.12 and run
pip-compile from inside it.

It also refactors dependency targets for add and upgrade separately.

Signed-off-by Chenxiong Qi <cqi@redhat.com>