feat: Work with migrations that use pmt-modify
STONEBLD-3800

This patch makes migrate command to work well during the transition to
where of all migrations have been written by using pmt-modify command.

TransitionToModifyCommandOperation class acts as a dispatcher to pass
pipeline file to apply_migrations method differently acccording to
whether all migrations, that belongs to a specific package file, use
pmt-modify. So, if not all of them use pmt-modify, use the legacy way to
handle pipeline file.

A log message is emitted for that detection. It might be used for
observing if it is time to disable/remove the dispatcher.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>