fix(yamleditor): Get line number of next sibling mapping field
STONEBLD-4024

For a given piece of YAML:

    - name: task2
      taskRef:
        params:
        - name: name
          value: task2
        - name: bundle
          value: bundle-ref
        resolver: bundles

The next sibling of node .taskRef.params.1 is mapping field resolver.
The problem is, method find_next_sibling returns string bundles as the
next sibling. As a result, calling of lc fails.

This patch fixes by calling lc.key() to return the line number of
mapping field directly. Changes are:

* Rename find_next_sibling to find_next_sibling_line and make it returns
  line number rather than a YAML node. No change to the meaning of
  return value None.
* Remove line number adjustment for CommentedMap.
* Argument node is in type CommentedSeq and CommentMap. isinstance call
  is updated to check the specific type. This makes type checker works,
  otherwise node.lc is reported as an error since dict does not have
  attribute lc.
* Extend test data to cover the problem.

Signed-off-by: Chenxiong Qi <cqi@redhat.com>